+++ modules/registrars/synergywholesaledomains/synergywholesaledomains.php
@@ -280,6 +280,33 @@ function synergywholesaledomains_getConfigArray(array $params)
             'Type' => 'yesno',
             'Description' => 'Enable this option to force update the WHOIS.json data<br><b>NOTE:</b> This option will be disabled automatically again once you have clicked \'Save Changes\' and the update sequence is completed.',
         ],
+        'defaultDnsConfig' => [
+            'FriendlyName' => 'Default DNS Config',
+            'Type' => 'dropdown',
+            'Options' => [
+                '0' => 'Nothing',
+                '1' => 'Nameservers',
+                '2' => 'FreeDNS with email forwarding',
+                '3' => 'Parked',
+                '4' => 'FreeDNS',
+                '5' => 'SWS Account Default',
+                '6' => 'Legacy Hosting',
+                '7' => 'Wholesale Hosting'
+            ],
+            'Description' => 'Which Default DNS Config will be applied to newly registered domains',
+        ],
+        'enableDnsManagement' => [
+            'FriendlyName' => 'Enable DNS Management',
+            'Type' => 'yesno',
+            'Size' => '1',
+            'Description' => 'Tick if you wish to enable DNS management on the domain, if Default DNS supports it.',
+        ],
+        'enableEmailForwarding' => [
+            'FriendlyName' => 'Enable Email Forwarding',
+            'Type' => 'yesno',
+            'Size' => '1',
+            'Description' => 'Tick if you wish to enable email forwarding on the domain, if Default DNS supports it.',
+        ],
         'Version' => [
             'Description' => 'This module version: ' . SW_MODULE_VERSION,
         ],
@@ -610,9 +637,10 @@ function synergywholesaledomains_RegisterDomain(array $params)
         $request['costPrice'] = $params['premiumCost'];
         $request['premium'] = true;
     }
-    
+
     try {
         synergywholesaledomains_apiRequest('domainRegister', $params, $request);
+
         return [
             'success' => true,
         ];
@@ -1511,7 +1539,7 @@ function synergywholesaledomains_DelURLForward(array $record, array $params)
 {
     return synergywholesaledomains_apiRequest('deleteSimpleURLForward', $params, [
         'recordID' => $record['record_id'],
-    ], $false);
+    ], false);
 }
 
 /**